requirements.txt
argon2-cffi==20.1.0
async-generator==1.10
attrs==20.2.0
backcall==0.2.0
beautifulsoup4==4.9.3
bleach==3.2.1
blis==0.7.4
cachetools==4.2.0
catalogue==1.0.0
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
ChatterBot==1.0.8
colorama==0.4.4
comtypes==1.1.7
cycler==0.10.0
cymem==2.0.5
decorator==4.4.2
defusedxml==0.6.0
en-core-web-sm==2.3.1
entrypoints==0.3
et-xmlfile==1.0.1
geographiclib==1.50
geopy==2.0.0
google-auth==1.24.0
google-auth-oauthlib==0.4.2
gspread==3.6.0
httplib2==0.18.1
ibm-cloud-sdk-core==1.7.3
ibm-watson==4.7.1
idna==2.10
importlib-metadata==2.0.0
ipykernel==5.3.4
ipython==7.18.1
ipython-genutils==0.2.0
ipywidgets==7.5.1
jdcal==1.4.1
jedi==0.17.2
Jinja2==2.11.2
joblib==0.17.0
json5==0.9.5
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.7
jupyter-console==6.2.0
jupyter-core==4.6.3
jupyterlab==2.2.8
jupyterlab-pygments==0.1.2
jupyterlab-server==1.2.0
keyboard==0.13.5
kiwisolver==1.3.1
MarkupSafe==1.1.1
mathparse==0.1.2
matplotlib==3.3.2
mistune==0.8.4
MouseInfo==0.1.3
murmurhash==1.0.5
nbclient==0.5.0
nbconvert==6.0.7
nbformat==5.0.8
nest-asyncio==1.4.1
notebook==6.1.4
numpy==1.19.2
oauth2client==4.1.3
oauthlib==3.1.0
opencv-contrib-python==4.4.0.46
opencv-python==4.4.0.44
openpyxl==3.0.5
packaging==20.4
pandas==1.1.3
pandocfilters==1.4.2
parso==0.7.1
pickleshare==0.7.5
Pillow==7.2.0
plac==1.1.3
preshed==3.0.5
prometheus-client==0.8.0
prompt-toolkit==3.0.8
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAutoGUI==0.9.52
pycparser==2.20
PyGetWindow==0.0.9
Pygments==2.7.1
PyJWT==1.7.1
PyMsgBox==1.0.9
pyparsing==2.4.7
pyperclip==1.8.1
pypiwin32==223
PyRect==0.1.4
pyrsistent==0.17.3
PyScreeze==0.1.26
python-dateutil==2.8.1
pyttsx3==2.90
PyTweening==1.0.3
pytz==2020.1
pywin32==228
pywinpty==0.5.7
pyzmq==19.0.2
qtconsole==4.7.7
QtPy==1.9.0
random2==1.0.1
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.6
scikit-learn==0.23.2
scipy==1.5.4
selenium==3.141.0
Send2Trash==1.5.0
six==1.15.0
sklearn==0.0
soupsieve==2.1
spacy==2.3.5
SpeechRecognition==3.8.1
SQLAlchemy==1.3.20
srsly==1.0.5
terminado==0.9.1
testpath==0.4.4
thinc==7.4.5
threadpoolctl==2.1.0
tornado==6.0.4
tqdm==4.54.1
traitlets==5.0.5
urllib3==1.25.10
wasabi==0.8.0
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==0.48.0
widgetsnbextension==3.5.1
wikipedia==1.4.0
XlsxWriter==1.3.7
zipp==3.3.0
